+++ src/main/java/com/storm/OutputTweet.java
@@ -1,4 +1,4 @@
-package com.microsoft.example;
+package com.storm;
 
 import java.util.List;
 import java.util.ArrayList;
