+++ src/main/java/com/storm/TwitterSpout.java
@@ -1,4 +1,4 @@
-package com.microsoft.example;
+package com.storm;
 
 import storm.trident.operation.TridentCollector;
